auth_server/requirements.txt
flask
pyopenssl
redis
python-dotenv